new changes

+++ Exepense_Vendor_Management/Models/EmployeeExpense.cs
@@ -12,8 +12,8 @@ namespace Expense_Vendor_Management.Models
         public bool isDeleted { get; set; }
         public string createdBy { get; set; }
         public DateTime submissionDate { get; set; }    
-        public float amount { get; set; }
-        public float? ApprovedAmount { get; set; }
+        public decimal amount { get; set; }
+        public decimal? ApprovedAmount { get; set; }
         public DateTime expenseOccurred { get; set; }
         public string expenseCategory { get; set; }
         public string Vid { get; set; }
